Refactory class location of the hive connector
